+++ native/App.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { StyleSheet } from 'react-native';
 import { NavigationContainer } from '@react-navigation/native'; //instalar
 import { createStackNavigator } from '@react-navigation/stack'; //instalar
@@ -8,13 +8,15 @@ import { st } from './store/store';
 import Estatistics from './screens/Estatistics';
 import Login from './screens/Login';
 import PosConsolidada from './screens/PosConsolidada';
-import Welcome from './screens/Welcome';
+import welcome from './screens/welcome';
 import SendMonyScreen from './screens/SendMonyScreen';
 import Transactions from './screens/Transactions';
 import RegisterModal from './components/RegisterModal';
 import AltaUSer from './screens/AltaUser';
 import FAQ from './screens/FAQ';
 import Toast from 'react-native-toast-message';
+import SideMenu from 'react-native-side-menu-updated';
+import MenuLateral from './screens/MenuLateral';
 
 // icons
 import Ionicons from 'react-native-vector-icons/Ionicons';
@@ -23,8 +25,8 @@ const Stack = createStackNavigator(); //contiene la navegacion
 
 function MainStack() {
 	return (
-		<Stack.Navigator initialRouteName='PosConsolidada'>
-			<Stack.Screen name='Welcome' component={Welcome} options={{ headerShown: false }} />
+		<Stack.Navigator>
+			<Stack.Screen name='Welcome' component={welcome} options={{ headerShown: false }} />
 
 			<Stack.Screen name='Login' component={Login} options={{ title: 'Iniciar sesión' }} options={{ headerShown: false }} />
 
@@ -55,19 +57,23 @@ function MainStack() {
 	);
 }
 export default function App() {
+	const menu = <MenuLateral></MenuLateral>;
+
 	return (
 		<Provider store={st}>
-			<NavigationContainer style={styles.container}>
-				<MainStack />
-				<Toast ref={(ref) => Toast.setRef(ref)} />
-			</NavigationContainer>
+			<SideMenu isOpen={false} autoClosing={true} menu={menu}>
+				<NavigationContainer style={styles.container}>
+					<MainStack />
+					<Toast ref={(ref) => Toast.setRef(ref)} />
+				</NavigationContainer>
+			</SideMenu>
 		</Provider>
 	);
 }
 const styles = StyleSheet.create({
 	container: {
 		flex: 1,
-		backgroundColor: '#fff',
+		backgroundColor: 'white',
 		alignItems: 'center',
 		justifyContent: 'center',
 	},
