+++ native/package.json
@@ -32,6 +32,7 @@
 		"react-native-reanimated": "~1.13.0",
 		"react-native-safe-area-context": "3.1.4",
 		"react-native-screens": "~2.10.1",
+		"react-native-side-menu-updated": "^1.2.1",
 		"react-native-svg": "^12.1.0",
 		"react-native-toast-message": "^1.3.4",
 		"react-native-vector-icons": "^7.1.0",
